Tweaks to the telemetry objects for the controller

+++ include/Telemetry/TelemetryObjects/Controller/Axis1Telemetry.h
@@ -1,11 +1,11 @@
 #include "Telemetry/TelemetryObjects/Telemetry.h"
 #include "vex.h"
 
-class Axis1Telemetry : public Telemetry<double> {
+class Axis1Telemetry : public Telemetry<int> {
 public:
 	Axis1Telemetry(vex::controller& c);
 
-	double newData();
+	int newData();
 
 private:
 	vex::controller& controllerRef;
