+++ include/Telemetry/TelemetryObjects/Controller/Axis2Telemetry.h
@@ -1,11 +1,11 @@
 #include "Telemetry/TelemetryObjects/Telemetry.h"
 #include "vex.h"
 
-class Axis2Telemetry : public Telemetry<double> {
+class Axis2Telemetry : public Telemetry<int> {
 public:
 	Axis2Telemetry(vex::controller& c);
 
-	double newData();
+	int newData();
 
 private:
 	vex::controller& controllerRef;
