+++ src/Telemetry/TelemetryObjects/Controller/Axis1Telemetry.cpp
@@ -1,14 +1,14 @@
 #include "Telemetry/TelemetryObjects/Controller/Axis1Telemetry.h"
 
-Axis1Telemetry::Axis1Telemetry(vex::controller& c) : Telemetry <double>(), controllerRef(c) {
-	previousData = TelemetryDataPoint<double>::create(
+Axis1Telemetry::Axis1Telemetry(vex::controller& c) : Telemetry <int>(), controllerRef(c) {
+	previousData = TelemetryDataPoint<int>::create(
 		0,
-		0.0,
+		0,
 		vex::timer::system(),
 		newData()
 	);
 }
 
-double Axis1Telemetry::newData() {
+int Axis1Telemetry::newData() {
 	return controllerRef.Axis1.position();
 }
